refactor: Consolidate duplicate audience-specific components
- Create unified components: JobEntry, Jobs, ProjectEntry, Projects, SkillEntry, Skills
- Remove 24 duplicate component files from audience-specific directories
- Remove 4 obsolete audience-specific page files
- Update CVPage to use new unified components
- Reduce codebase from 30+ duplicate files to 6 shared components
