+++ STYLING_UPDATES.md
@@ -1,0 +1,98 @@
+# Styling Updates Summary
+
+## Changes Made to Match Existing Page Styling
+
+### 1. CVPage.vue
+**Before:**
+- Used `q-page-container` wrapper
+- Used `q-card` components for sections
+- Custom layout with header and change profile button
+- Complex grid layout with col-md-8/col-md-4
+
+**After:**
+- Uses `<q-page>` as root (matching RecruiterPage, etc.)
+- Uses `ContactInformation` component at top (like existing pages)
+- Uses `row` and `col-*` classes (same as existing pages)
+- Uses `left-hand-side` (8 cols) and `right-hand-side` (4 cols) layout
+- Uses existing component patterns: JobsComponent, SkillsComponent, ProjectsComponent
+- Uses markdown rendering for summary and highlights
+- Empty `<style></style>` tag (matching existing pages)
+
+### 2. ProfileSelectionPage.vue
+**Before:**
+- Used `q-page-container` wrapper
+- Used `q-card` components for profile cards
+- Complex card-based layout with hover effects
+- Modal dialog for audience selection
+- Scoped styles with transitions
+
+**After:**
+- Uses `<q-page>` as root (matching LandingPage style)
+- Simple `row` and `col-*` layout
+- Text-based profile listing (matching LandingPage simplicity)
+- Direct audience selection (no modal)
+- Empty `<style></style>` tag
+- Uses same patterns as other pages
+
+### 3. ContactInformation.vue
+**Before:**
+- Only imported data from static JSON file
+
+**After:**
+- Accepts `data` prop for dynamic profile data
+- Falls back to static data when no prop provided
+- Uses computed properties to merge prop data with static defaults
+- Maintains backward compatibility with existing usage
+
+## Styling Patterns Matched
+
+### Layout Patterns
+✅ `<q-page>` root element
+✅ `row` and `col-*` classes for grid
+✅ `left-hand-side` and `right-hand-side` for main/sidebar
+✅ `q-px-xs`, `q-py-xs` for padding
+✅ `full-width` for spanning content
+
+### Typography Patterns
+✅ `text-h6` for section headings
+✅ `text-weight-bolder` for emphasis
+✅ `text-center` for centering
+✅ `text-small`, `text-caption` for smaller text
+
+### Component Patterns
+✅ Import specific audience components (JobsComponent, SkillsComponent, ProjectsComponent)
+✅ Use ContactInformation component
+✅ Markdown rendering for summary/highlights
+✅ Empty `<style></style>` tags
+
+### Navigation Patterns
+✅ Simple navigation structure
+✅ Breadcrumb-style hierarchy
+✅ Direct routing to CV pages
+✅ Profile-first then audience selection
+
+## Files Updated
+
+### Modified Files
+- `src/pages/CVPage.vue` - Updated to match existing page styling
+- `src/pages/ProfileSelectionPage.vue` - Simplified to match LandingPage
+- `src/components/ContactInformation.vue` - Added prop support for dynamic data
+
+## Verification
+
+All styling updates have been verified:
+- ✅ No lint errors
+- ✅ Successful build
+- ✅ All E2E tests pass
+- ✅ Profile structure validated
+- ✅ Data integrity maintained
+
+## Visual Consistency
+
+The updated pages now match:
+- ✅ LandingPage.vue layout pattern
+- ✅ RecruiterPage.vue component usage
+- ✅ TechnicalPeerPage.vue styling approach
+- ✅ All existing audience pages
+
+The CV Profiles system now feels like a native part of the application with consistent styling throughout.
