+++ rules/task-instructions.md
@@ -1,0 +1,8 @@
+Use `process-task-list` and the memory-loaded language rules to generate atomic tasks for: <feature/PRD summary>
+
+  - Reference the codebase before task generation
+  - Embed Go guidelines and rules inline where applicable
+  - Use one expectation per `it`, Red→Green→Refactor model
+  - Follow existing patterns and avoid side effects
+  - Return checklist only under `## Tasks`
+  - Do not explain or regenerate later
