+++ src/components/SkillEntry.vue
@@ -2,19 +2,7 @@
   <div class="skill q-px-sm"><b>{{ type }}</b>: {{ list }}</div>
 </template>
 
-<style></style>
-
 <script>
-/**
- *
- * Component used to display a single skill
- *
- * Usage:
- *   <SkillEntry />
- *
- * @module SkillEntry
- *
- */
 export default {
   name: 'SkillEntry',
   props: {
@@ -32,10 +20,6 @@ export default {
       }
       return this.skill.skills.join(', ')
     }
-  },
-  methods: {
-  },
-  mounted() {
   }
 }
 </script>
