+++ src/components/Skills.vue
@@ -2,33 +2,14 @@
   <div id="skills">
     <div class="text-h6 col row text-weight-bolder">Skills</div>
     <div id="skills-list" class="flex row wrap justify-between">
-      <SkillEntry v-for="skill in this.sortedSkills" :key="skill" :skill="skill" />
+      <SkillEntry v-for="skill in this.sortedSkills" :key="skill.type" :skill="skill" />
     </div>
   </div>
 </template>
 
-<style>
-#skills-list {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-  width: 100%;
-}
-</style>
-
 <script>
-import SkillEntry from '@recruiter/Skill.vue'
+import SkillEntry from './SkillEntry.vue'
 
-/**
- *
- * Component used to display the skills section
- *
- * Usage:
- *   <SkillsComponent :skills="this.cv.skills" />
- *
- * @module SkillsComponent
- *
- */
 export default {
   name: 'SkillsComponent',
   props: {
@@ -45,10 +26,17 @@ export default {
       const skills = Object.keys(this.skills).map(key => {
         return { type: key, skills: this.skills[key] }
       })
-
       return skills
     }
   }
-
 }
 </script>
+
+<style>
+#skills-list {
+  list-style-type: none;
+  padding: 0;
+  margin: 0;
+  width: 100%;
+}
+</style>
